# RavenDB-19730 - Sharding - Query - Further Performance Optimizations