--- a/src/Raven.Server/Config/Settings/TimeSetting.cs
+++ b/src/Raven.Server/Config/Settings/TimeSetting.cs
@@ -45,6 +45,11 @@ namespace Raven.Server.Config.Settings
             }
         }
 
+        public TimeSetting(TimeSpan timeSpan)
+        {
+            AsTimeSpan = timeSpan;
+        }
+
         internal long GetValue(TimeUnit requestedUnit)
         {
             switch (requestedUnit)