--- a/src/Raven.Server/Documents/Sharding/Commands/Querying/AbstractShardedQueryCommand.cs
+++ b/src/Raven.Server/Documents/Sharding/Commands/Querying/AbstractShardedQueryCommand.cs
@@ -1,4 +1,5 @@
-﻿using System.Net.Http;
+﻿using System;
+using System.Net.Http;
 using System.Text;
 using Raven.Client.Documents.Commands;
 using Raven.Client.Documents.Queries;
@@ -25,8 +26,9 @@ public abstract class AbstractShardedQueryCommand<TResult, TParameters> : Abstra
         bool ignoreLimit,
         string indexName,
         bool canReadFromCache,
-        string raftUniqueRequestId)
-        : base(indexQuery, true, metadataOnly, indexEntriesOnly, ignoreLimit)
+        string raftUniqueRequestId,
+        TimeSpan globalHttpClientTimeout)
+        : base(indexQuery, true, metadataOnly, indexEntriesOnly, ignoreLimit, globalHttpClientTimeout)
     {
         _query = query;
         Scope = scope;
@@ -51,6 +53,6 @@ public abstract class AbstractShardedQueryCommand<TResult, TParameters> : Abstra
     {
         DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Grisha, DevelopmentHelper.Severity.Normal, "let's create a server-side query class here and use same code as for QueryCommand");
 
-        return new StringContent(_query.ToString(), Encoding.UTF8, "application/json");
+        return new StringContent(_query, Encoding.UTF8, "application/json");
     }
 }