… src/Raven.Server/Documents/Sharding/Operations/IShardedOperation.cs
… src/Raven.Server/Documents/Sharding/Operations/Queries/ShardedQueryOperation.cs
… src/Raven.Server/Documents/Sharding/Queries/AbstractShardedQueryProcessor.cs
… src/Raven.Server/Documents/Sharding/Queries/Facets/ShardedFacetedQueryProcessor.cs
… src/Raven.Server/Documents/Sharding/Queries/IndexEntries/ShardedIndexEntriesQueryProcessor.cs
… src/Raven.Server/Documents/Sharding/Queries/ShardedQueryProcessorBase.cs
… src/Raven.Server/Documents/Sharding/Queries/ShardedQueryStreamProcessor.cs
… src/Raven.Server/Documents/Sharding/Queries/Suggestions/ShardedSuggestionQueryProcessor.cs