### Issue link

https://issues.hibernatingrhinos.com/issue/RavenDB-19730

### Additional description

Improvements made:

- Configurable `GZip` compression between orchestrator and shards - enable/disable. Spent a lot of time on compressing and decompressing. default is no compression;

- Configurable `GlobalHttpClientTimeout` for orchestrator - spent a long time in Microsoft's `HttpClient`'s `CancelAfter()` because there was contention on a lock inside. This function was only called when there was a timeout set for the `HttpClient`.  In order to avoid this path, the timeout configuration's default is set to infinite (-1 ms).
(see here https://github.com/dotnet/runtime/blob/54972b75762f800fddbe1abfe5fce544ae671033/src/libraries/System.Net.Http/src/System/Net/Http/HttpClient.cs#L806").

- Avoid cloning the blittable results that exist on each shard's context to the local orchestrator context. Now keeping them alive until they are written to the stream and only then disposing.

- Now converting the query into a string (`Blittable.ToString()` function) so it will happen once if possible instead of per shard. When the request is not load by id, it is the same for all shards.

Results:

Running the following queries:
Search: `from "Questions" where search("Title",$p0) select Title limit 128`
Load: `session.LoadAsync<dynamic>(someId)`

Client keeps sending requests to maintain a set number of concurrent requests during the entire run.
Run length is ~30 seconds.

**Search Queries**
All queries are for searching on terms that return the most results.
Concurrency: 10

Before Optimizations:
<html>
<body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Sharding | 117 | 83.52 | 154 | 184 | 1421 | 15 | 1447 | 3540
NonSharded | 246 | 39 | 83 | 106 | 1311 | 16 | 1381 | 7412
Orch seperated | 209 | 46.3 | 103 | 134 | 1350 | 17 | 1413 | 6302

<!--EndFragment-->
</body>
</html>

After Optimizations:
<html>
<body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Sharding | 163 | 59.78 | 143 | 184 | 1189 | 20 | 1261 | 4908
NonSharded | 227 | 42.35 | 101 | 122 | 1510 | 13 | 1562 | 6841
Orch seperated | 221 | 43.66 | 101 | 132 | 1592 | 20 | 1635 | 6653

<!--EndFragment-->
</body>
</html>

After Rebase:
<html>
<body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Sharding | 173 | 56.13 | 108 | 123 | 1286 | 15 | 1320 | 5229
NonSharded | 245 | 39.31 | 97 | 109 | 1330 | 10 | 1409 | 7369
Orch seperated | 235 | 40.98 | 91 | 128 | 1208 | 21 | 1266 | 7066

<!--EndFragment-->
</body>
</html>

**Load Queries**
All queries are for Ids that exist in the database
Concurrency: 50
(numbers are a bit unstable)

Before optimizations:
<html>
<body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Sharding | 1971 | 23.53 | 48 | 61 | 105 | 50 | 2689 | 59186
NonSharded | 3436 | 10.15 | 29 | 36 | 60 | 51 | 2748 | 103120
Orch seperated | 2237 | 19.77 | 46 | 69 | 109 | 53 | 2582 | 67160

<!--EndFragment-->
</body>
</html>

After optimizations:
<html>
<body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Sharding | 2777 | 15.11 | 38 | 51 | 175 | 52 | 2624 | 83375
NonSharded | 3267 | 11.51 | 32 | 36 | 63 | 52 | 2436 | 98031
Orch seperated | 2755 | 15.37 | 42 | 52 | 501 | 111 | 1978 | 82687

<!--EndFragment-->
</body>
</html>

After Rebase:
<html>
<body>
<!--StartFragment--><google-sheets-html-origin>

  | req /s | avg | p99 | p995 | p999 | >200ms | max | total reqs
-- | -- | -- | -- | -- | -- | -- | -- | --
Sharding | 2100 | 21.96 | 48 | 68 | 176 | 51 | 2514 | 63056
NonSharded | 3305 | 10.83 | 31 | 39 | 67 | 58 | 2100 | 99185
Orch seperated | 2448 | 18.62 | 55 | 66 | 100 | 52 | 1787 | 73472

<!--EndFragment-->
</body>
</html>

### Type of change

- Optimization

### How risky is the change?

- Moderate 

### Backward compatibility

- Non breaking change

### Documentation update

- This change requires a documentation update.

### Testing by Contributor

- It has been verified by manual testing

### Testing by RavenDB QA team

- This change requires a special QA testing due to possible performance or resources usage implications (CPU, memory, IO). Please mark the issue on YouTrack using `QA Required` tag.

### UI work

- It requires further work in the Studio - changing timeout configuration to infinite should be possible through studio