@ayende 2023-02-21
Why do we have a drop on the non sharded scenario after optimizations?