@arekpalinski 2023-02-27
I'm merging it. We'll continue the perf analysis after we run some QA benchmarks.